Add files via upload
Data of covid-19 is huge in includes number of deaths, recovered people, new cases. In this paper shows the text classification of covid-19 dataset over the world as well as in United States.
This data is collected from Kaggle and used for text mining analysis. The data set contain 628319 metadata rows.

This paper analyzes the following questions.
Q1.	Show the number of confirmed deaths and recovered cases in USA.
Q2.	Show the number of confirmed Deaths and Recovered cases in world
Q3.	Remove all the records where the confirmed case is less than 10,000
Q4.	The place where the maximum number of cases are recorded 
Q5.	The place where the minimum number of death cases were recorded.
Q6.	Sort the entire data with respect to the confirmed cases in ascending order
Q7.	Sort the entire data with respect to number of recovered cases in descending order